content_forge/requirements.txt
# AWS SDK and Services
boto3==1.34.144
botocore==1.34.144

# Data Validation and Serialization
pydantic==2.8.2
pydantic-settings==2.3.4

# AWS Lambda Utilities
aws-lambda-powertools==2.42.0

# Authentication and Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
cryptography==42.0.8

# HTTP Client
httpx==0.27.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
hypothesis==6.108.1
moto==5.0.9

# Utilities
python-dotenv==1.0.1
uuid==1.30
typing-extensions==4.12.2

# Frontend (Streamlit)
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
requests==2.32.3
